# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Ticket Booking System",
        "title": "Ticket Booking System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает список всех пользователей системы с фильтрацией и пагинацией. / [EN] Returns a list of all system users with filtering and pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "[RU] Список всех пользователей / [EN] All users list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "[RU] Страница / [EN] Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "[RU] Лимит / [EN] Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "[RU] Статус / [EN] Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "[RU] Роль / [EN] Role",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает статистику по количеству и статусам пользователей. / [EN] Returns statistics on user count and statuses.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "[RU] Статистика пользователей / [EN] User stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.UserStats"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает полную информацию о пользователе, включая его историю. / [EN] Returns full user information including history.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "[RU] Детальная информация о пользователе / [EN] User details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID пользователя / [EN] User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.UserDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/block": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Переводит статус пользователя в BLOCKED. / [EN] Changes user status to BLOCKED.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "[RU] Заблокировать пользователя / [EN] Block user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID пользователя / [EN] User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/role": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Позволяет администратору изменить роль пользователя. / [EN] Allows admin to change user role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "[RU] Изменить роль пользователя / [EN] Update user role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID пользователя / [EN] User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "[RU] Новая роль / [EN] New role",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "role": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/unblock": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Восстанавливает активный статус пользователя. / [EN] Restores active user status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "[RU] Разблокировать пользователя / [EN] Unblock user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID пользователя / [EN] User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.Role": {
            "type": "string",
            "enum": [
                "guest",
                "user",
                "organizer",
                "moderator",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleGuest",
                "RoleUser",
                "RoleOrganizer",
                "RoleModerator",
                "RoleAdmin"
            ]
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.Status": {
            "type": "string",
            "enum": [
                "active",
                "blocked",
                "pending_verification"
            ],
            "x-enum-varnames": [
                "StatusActive",
                "StatusBlocked",
                "StatusPendingVerification"
            ]
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Role"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Status"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.UserDetail": {
            "type": "object",
            "properties": {
                "statistics": {
                    "type": "object",
                    "additionalProperties": true
                },
                "user": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.User"
                }
            }
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.UserStats": {
            "type": "object",
            "properties": {
                "active_users": {
                    "type": "integer"
                },
                "blocked_users": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                },
                "unverified_users": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}